refactor: split catalog.rs into per-provider modules
1,396-line catalog.rs → catalog/ directory. Each provider parser
(openrouter, groq, nvidia, anthropic, codex, generic) gets its
own file. Shared types, trait, fetch functions, and tests stay
in mod.rs. All 45 tests pass. No API changes.
