+++ src/runtime/openai/catalog/mod.rs
@@ -26,19 +26,26 @@ use std::collections::BTreeMap;
 use std::future::Future;
 use std::pin::Pin;
 use std::time::Duration;
-use serde::Deserialize;
 
 pub const CATALOG_REQUEST_TIMEOUT: Duration = Duration::from_secs(15);
-const ANTHROPIC_MODELS_URL: &str = "https://api.anthropic.com/v1/models";
-const ANTHROPIC_MODELS_PAGE_LIMIT: usize = 100;
 const ANTHROPIC_MODELS_MAX_PAGES: usize = 20;
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AnthropicCatalogPage {
-    pub models: Vec<CatalogModel>,
-    pub has_more: bool,
-    pub last_id: Option<String>,
-}
+mod anthropic;
+mod codex;
+mod generic;
+mod groq;
+mod nvidia;
+mod openrouter;
+
+pub use anthropic::{
+    anthropic_models_url, merge_catalog_pages, parse_anthropic_catalog_models,
+    parse_anthropic_catalog_page, AnthropicCatalogPage,
+};
+pub use codex::codex_static_catalog_models;
+pub use generic::parse_generic_catalog_models;
+pub use groq::{infer_groq_reasoning, parse_groq_catalog_models};
+pub use nvidia::{infer_nvidia_reasoning, parse_nvidia_catalog_models};
+pub use openrouter::parse_openrouter_catalog_models;
 
 // ─── Modality ────────────────────────────────────────────────────────────────
 
@@ -238,376 +245,6 @@ pub fn static_seeds_from_spec(
             from_static_seed(spec.key, spec.name, id, label)
         })
         .collect()
-}
-
-// ─── Generic OpenAI-compatible parser ────────────────────────────────────────
-
-#[derive(Debug, Deserialize)]
-struct GenericModelsResponse {
-    data: Vec<GenericModelItem>,
-}
-
-#[derive(Debug, Deserialize)]
-struct GenericModelItem {
-    id: String,
-    #[serde(default)]
-    name: Option<String>,
-}
-
-/// Parse a `{data:[{id,name?}]}` response.
-/// Identical filtering behaviour to `registry::parse_provider_models_response`.
-pub fn parse_generic_catalog_models(
-    body: &str,
-    provider_key: &str,
-    provider_name: &str,
-) -> Result<Vec<CatalogModel>, serde_json::Error> {
-    let resp: GenericModelsResponse = serde_json::from_str(body)?;
-    Ok(resp
-        .data
-        .into_iter()
-        .filter_map(|item| {
-            let mut m = CatalogModel::new(provider_key, provider_name, item.id)?;
-            m.label = item.name.filter(|n| !n.trim().is_empty());
-            m.reasoning = ReasoningSupport::GenericOpenAi;
-            Some(m)
-        })
-        .collect())
-}
-
-// ─── OpenRouter parser ────────────────────────────────────────────────────────
-
-#[derive(Debug, Deserialize)]
-struct OpenRouterModelsResponse {
-    data: Vec<OpenRouterModelItem>,
-}
-
-#[derive(Debug, Deserialize)]
-struct OpenRouterModelItem {
-    id: String,
-    #[serde(default)]
-    name: Option<String>,
-    #[serde(default)]
-    context_length: Option<u64>,
-    #[serde(default)]
-    architecture: Option<OpenRouterArchitecture>,
-    #[serde(default)]
-    pricing: Option<OpenRouterPricing>,
-    #[serde(default)]
-    top_provider: Option<OpenRouterTopProvider>,
-    #[serde(default)]
-    supported_parameters: Vec<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct OpenRouterArchitecture {
-    #[serde(default)]
-    input_modalities: Vec<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct OpenRouterPricing {
-    #[serde(default)]
-    prompt: Option<String>,
-    #[serde(default)]
-    completion: Option<String>,
-    #[serde(default)]
-    internal_reasoning: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct OpenRouterTopProvider {
-    #[serde(default)]
-    max_completion_tokens: Option<u64>,
-}
-
-/// Parse an OpenRouter `/models` JSON body into CatalogModel entries.
-/// Pure function — no network I/O — safe to unit-test with fixtures.
-pub fn parse_openrouter_catalog_models(body: &str) -> Result<Vec<CatalogModel>, serde_json::Error> {
-    let resp: OpenRouterModelsResponse = serde_json::from_str(body)?;
-    Ok(resp
-        .data
-        .into_iter()
-        .filter_map(|item| {
-            let mut m = CatalogModel::new("openrouter", "OpenRouter", item.id)?;
-            m.provider_kind = CatalogProviderKind::OpenRouter;
-            m.label = item.name.filter(|n| !n.trim().is_empty());
-            m.context_tokens = item.context_length;
-            m.max_output_tokens = item.top_provider.as_ref().and_then(|p| p.max_completion_tokens);
-
-            let mods: Vec<Modality> = item
-                .architecture
-                .map(|a| a.input_modalities.iter().map(|s| Modality::from_str(s)).collect())
-                .unwrap_or_else(|| vec![Modality::Text]);
-            m.input_modalities = mods;
-
-            let pricing = item.pricing.map(|p| PricingSummary {
-                prompt:             p.prompt.filter(|v| !v.trim().is_empty()),
-                completion:         p.completion.filter(|v| !v.trim().is_empty()),
-                internal_reasoning: p.internal_reasoning.filter(|v| !v.trim().is_empty()),
-            }).unwrap_or_default();
-
-            let has = |param: &str| item.supported_parameters.iter().any(|p| p == param);
-            let internal_priced = pricing.has_internal_reasoning_cost();
-
-            m.reasoning = if has("verbosity") {
-                // Anthropic models routed through OpenRouter
-                ReasoningSupport::AnthropicAdaptive { adaptive: true }
-            } else if has("reasoning") || has("include_reasoning") || has("reasoning_effort") || internal_priced {
-                ReasoningSupport::OpenRouter {
-                    include_reasoning: has("include_reasoning"),
-                    effort:            has("reasoning_effort"),
-                    verbosity:         false,
-                    internal_reasoning_priced: internal_priced,
-                }
-            } else {
-                ReasoningSupport::None
-            };
-
-            m.pricing = pricing;
-            m.source = CatalogSource::Live;
-            Some(m)
-        })
-        .collect())
-}
-
-
-// ─── Groq parser / inference ────────────────────────────────────────────────
-
-#[derive(Debug, Deserialize)]
-struct GroqModelsResponse {
-    data: Vec<GroqModelItem>,
-}
-
-#[derive(Debug, Deserialize)]
-struct GroqModelItem {
-    id: String,
-    #[serde(default = "default_true")]
-    active: bool,
-    #[serde(default)]
-    context_window: Option<u64>,
-    #[serde(default)]
-    owned_by: Option<String>,
-}
-
-fn default_true() -> bool { true }
-
-pub fn infer_groq_reasoning(model_id: &str) -> ReasoningSupport {
-    let id = model_id.to_ascii_lowercase();
-    if id.starts_with("openai/gpt-oss-") || id.starts_with("qwen/qwen3-") || id.starts_with("groq/compound") {
-        ReasoningSupport::GroqReasoning
-    } else {
-        ReasoningSupport::None
-    }
-}
-
-pub fn parse_groq_catalog_models(body: &str) -> Result<Vec<CatalogModel>, serde_json::Error> {
-    let resp: GroqModelsResponse = serde_json::from_str(body)?;
-    Ok(resp
-        .data
-        .into_iter()
-        .filter(|item| item.active)
-        .filter_map(|item| {
-            let mut m = CatalogModel::new("groq", "Groq", item.id)?;
-            m.provider_kind = CatalogProviderKind::Groq;
-            m.label = item.owned_by.as_ref().map(|owner| format!("{} — {}", m.id, owner));
-            m.context_tokens = item.context_window;
-            m.reasoning = infer_groq_reasoning(&m.id);
-            m.source = CatalogSource::Live;
-            Some(m)
-        })
-        .collect())
-}
-
-// ─── NVIDIA NIM parser / inference ──────────────────────────────────────────
-
-#[derive(Debug, Deserialize)]
-struct NvidiaModelsResponse {
-    data: Vec<NvidiaModelItem>,
-}
-
-#[derive(Debug, Deserialize)]
-struct NvidiaModelItem {
-    id: String,
-    #[serde(default)]
-    owned_by: Option<String>,
-}
-
-fn infer_nvidia_context_tokens(model_id: &str) -> Option<u64> {
-    let id = model_id.to_ascii_lowercase();
-    if id.contains("kimi-k2-thinking") { Some(256_000) }
-    else if id.contains("nemotron-ultra")
-        || id.contains("nemotron-super")
-        || id.contains("qwen3-next")
-        || id.contains("deepseek-v3.1")
-        || id.contains("llama-3.1-405b")
-        || id.contains("llama-3.2-90b-vision")
-        || id.contains("mistral-large-2")
-    { Some(128_000) }
-    else { None }
-}
-
-pub fn infer_nvidia_reasoning(model_id: &str) -> ReasoningSupport {
-    let id = model_id.to_ascii_lowercase();
-    if id.contains("thinking")
-        || id.contains("cosmos-reason")
-        || id.contains("nemotron-ultra")
-        || id.contains("nemotron-super")
-        || id.contains("mistral-nemotron")
-        || id.contains("magistral")
-        || id.contains("deepseek-v4")
-        || id.contains("kimi")
-        || id.contains("reasoning")
-    {
-        ReasoningSupport::NvidiaInlineThinking
-    } else {
-        ReasoningSupport::None
-    }
-}
-
-pub fn parse_nvidia_catalog_models(body: &str) -> Result<Vec<CatalogModel>, serde_json::Error> {
-    let resp: NvidiaModelsResponse = serde_json::from_str(body)?;
-    let mut seen = std::collections::BTreeSet::new();
-    Ok(resp
-        .data
-        .into_iter()
-        .filter_map(|item| {
-            let id = item.id;
-            if id.trim().is_empty() || !seen.insert(id.clone()) {
-                return None;
-            }
-            let mut m = CatalogModel::new("nvidia", "NVIDIA NIM", id)?;
-            m.provider_kind = CatalogProviderKind::NvidiaNim;
-            m.label = item.owned_by.as_ref().map(|owner| format!("{} — {}", m.id, owner));
-            m.context_tokens = infer_nvidia_context_tokens(&m.id);
-            m.reasoning = infer_nvidia_reasoning(&m.id);
-            m.source = if m.context_tokens.is_some() || m.reasoning != ReasoningSupport::None {
-                CatalogSource::Inferred
-            } else {
-                CatalogSource::Live
-            };
-            Some(m)
-        })
-        .collect())
-}
-
-
-// ─── Anthropic parser and Codex static handler ──────────────────────────────
-
-#[derive(Debug, Deserialize)]
-struct AnthropicModelsPage {
-    data: Vec<AnthropicModelItem>,
-    #[serde(default)]
-    has_more: bool,
-    #[serde(default)]
-    last_id: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct AnthropicModelItem {
-    id: String,
-    #[serde(default)]
-    display_name: Option<String>,
-    #[serde(default)]
-    max_input_tokens: Option<u64>,
-    #[serde(default)]
-    max_tokens: Option<u64>,
-    #[serde(default)]
-    capabilities: Option<AnthropicCapabilities>,
-}
-
-#[derive(Debug, Deserialize)]
-struct AnthropicCapabilities {
-    #[serde(default)]
-    thinking: Option<CapabilitySupported>,
-    #[serde(default)]
-    effort: Option<AnthropicEffortCapability>,
-}
-
-#[derive(Debug, Deserialize)]
-struct CapabilitySupported {
-    #[serde(default)]
-    supported: bool,
-}
-
-#[derive(Debug, Deserialize)]
-struct AnthropicEffortCapability {
-    #[serde(default)]
-    supported: bool,
-}
-
-pub fn parse_anthropic_catalog_page(body: &str) -> Result<AnthropicCatalogPage, serde_json::Error> {
-    let page: AnthropicModelsPage = serde_json::from_str(body)?;
-    let models = page
-        .data
-        .into_iter()
-        .filter_map(|item| {
-            let mut m = CatalogModel::new("anthropic", "Anthropic", item.id)?;
-            m.provider_kind = CatalogProviderKind::Anthropic;
-            m.label = item.display_name.filter(|name| !name.trim().is_empty());
-            m.context_tokens = item.max_input_tokens;
-            m.max_output_tokens = item.max_tokens;
-            m.reasoning = match item.capabilities {
-                Some(caps) if caps.thinking.as_ref().is_some_and(|c| c.supported) => {
-                    ReasoningSupport::AnthropicAdaptive {
-                        adaptive: caps.effort.as_ref().is_some_and(|c| c.supported),
-                    }
-                }
-                _ => ReasoningSupport::Unknown,
-            };
-            m.source = CatalogSource::Live;
-            Some(m)
-        })
-        .collect();
-
-    Ok(AnthropicCatalogPage {
-        models,
-        has_more: page.has_more,
-        last_id: page.last_id.filter(|id| !id.trim().is_empty()),
-    })
-}
-
-pub fn parse_anthropic_catalog_models(body: &str) -> Result<Vec<CatalogModel>, serde_json::Error> {
-    parse_anthropic_catalog_page(body).map(|page| page.models)
-}
-
-pub fn anthropic_models_url(after_id: Option<&str>) -> String {
-    let mut url = format!("{ANTHROPIC_MODELS_URL}?limit={ANTHROPIC_MODELS_PAGE_LIMIT}");
-    if let Some(after_id) = after_id.filter(|id| !id.trim().is_empty()) {
-        url.push_str("&after_id=");
-        url.push_str(after_id);
-    }
-    url
-}
-
-pub fn merge_catalog_pages(pages: Vec<Vec<CatalogModel>>) -> Vec<CatalogModel> {
-    let mut seen = std::collections::BTreeSet::new();
-    let mut merged = Vec::new();
-    for page in pages {
-        for model in page {
-            if seen.insert(model.id.clone()) {
-                merged.push(model);
-            }
-        }
-    }
-    merged
-}
-
-pub fn codex_static_catalog_models() -> Vec<CatalogModel> {
-    [
-        ("gpt-5.5", "GPT-5.5"),
-        ("gpt-5.4", "GPT-5.4"),
-        ("gpt-5.4-mini", "GPT-5.4 Mini"),
-    ]
-    .into_iter()
-    .filter_map(|(id, label)| {
-        let mut m = CatalogModel::new("openai-codex", "OpenAI Codex", id)?;
-        m.provider_kind = CatalogProviderKind::OpenAiCodex;
-        m.label = Some(label.to_string());
-        m.reasoning = ReasoningSupport::Unknown;
-        m.source = CatalogSource::StaticFallback;
-        Some(m)
-    })
-    .collect()
 }
 
 // ─── Live fetch helpers ───────────────────────────────────────────────────────
@@ -1168,7 +805,6 @@ mod tests {
     }
 
     // ── Task 3: Generic handler / compat with registry ────────────────────────
-
